+++ src/main/java/it/eng/parer/crypto/web/config/security/StrictSecurityConfig.java
@@ -22,7 +22,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 
 /*
  * since spring boot 3.x
@@ -41,11 +40,11 @@ public class StrictSecurityConfig {
 
 	http.csrf(csrf -> csrf.disable()) // disable csrf
 		.authorizeHttpRequests(authorizeHttpRequests -> authorizeHttpRequests
-			.requestMatchers(new AntPathRequestMatcher(URL_ADMIN_BASE + RESOURCE_INFOS))
+			.requestMatchers(URL_ADMIN_BASE + RESOURCE_INFOS)
 			.authenticated())
 		.httpBasic(withDefaults()) // basic auth
 		.authorizeHttpRequests(authorizeHttpRequests -> authorizeHttpRequests
-			.requestMatchers(new AntPathRequestMatcher(URL_ADMIN_BASE + "/**"))
+			.requestMatchers(URL_ADMIN_BASE + "/**")
 			.denyAll()) // deny admin
 		.authorizeHttpRequests(auth -> auth.anyRequest().permitAll()); // permit all