+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## 1.22.0 (18-09-2025)
+
+### Novità: 1
+- [#39102](https://parermine.regione.emilia-romagna.it/issues/39102)  Modifica servizio estrazione documento originale da P7m
+
 ## 1.21.0 (07-05-2025)
 
 ### Novità: 1