+++ RELEASE-NOTES.md
@@ -1,4 +1,4 @@
-## 1.21.0 (07-05-2025)
+## 1.22.0 (18-09-2025)
 
 ### Novità: 1
-- [#37615](https://parermine.regione.emilia-romagna.it/issues/37615) Aggiornamento librerie obsolete primo quadrimestre 2025
+- [#39102](https://parermine.regione.emilia-romagna.it/issues/39102)  Modifica servizio estrazione documento originale da P7m